requirements.txt
accelerate==0.25.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
blinker==1.7.0
boto3==1.35.90
botocore==1.35.90
certifi==2023.11.17
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
crispy-bootstrap5==2024.2
cryptography==43.0.1
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
django-js-asset==2.2.0
django-markdownx==4.0.7
django-reversion==5.1.0
django-storages==1.14.4
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.53.0
fsspec==2023.12.2
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.20.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.25.2
packaging==23.2
pandas==2.1.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
platformdirs==4.3.6
priority==1.3.0
prompt_toolkit==3.0.48
psutil==5.9.7
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==308
PyYAML==6.0.1
pyzmq==26.2.0
redis==5.1.0
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.4.1
scikit-learn==1.3.0
scipy==1.11.2
service-identity==24.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tinycss2==1.4.0
tokenizers==0.15.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.36.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
yarg==0.1.9
zope.interface==7.0.3
